feat: update app name and colors in manifest, enhance MembersChecked date formatting, and add click outside functionality to QrCodeButton

+++ src/components/AdminDashboard/AdminRouterPage/MembersChecked.jsx
@@ -148,9 +148,9 @@ const MembersChecked = ( { changeColor }) => {
                                             "Today"
                                         ) : (
                                             new Date(date).toLocaleDateString("en-GB", {
-                                                weekday: 'long',
+                                                weekday: 'short',
                                                 day: 'numeric',
-                                                month: 'short',
+                                                month: 'long',
                                                 year: 'numeric',
                                             })
                                         )
@@ -206,7 +206,7 @@ const MembersChecked = ( { changeColor }) => {
                                                 new Date(date).toLocaleDateString("en-GB", {
                                                     weekday: 'long',
                                                     day: 'numeric',
-                                                    month: 'short',
+                                                    month: 'long',
                                                     year: 'numeric',
                                                 })
                                             )
